fixing drop functions

+++ foundation-database/public/functions/createarcreditmemo.sql
@@ -1,4 +1,4 @@
-SELECT DROP FUNCTION IF EXISTS createarcreditmemo(integer, integer, text, text, date, numeric, text, integer, integer, integer, date, integer, integer, numeric, integer, integer, integer, integer, numeric);
+SELECT DROP FUNCTION IF EXISTS createarcreditmemo(integer, integer, text, text, date, numeric, text, integer, integer, integer, date, integer, integer, numeric, integer, integer, integer, integer, integer);
 
 CREATE OR REPLACE FUNCTION createARCreditMemo(pId            INTEGER,
                                               pCustid        INTEGER,
